+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added hasThumbnail and disabled properties to the Tile component. 
 - Added as hover image preview icon for the Tile component.
 - Adding the `tab` component.
+- Adding props for error handling in Preview component.
 
 ### Breaking changes